backend/requirements.txt
# Core FastAPI framework and server
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# WhatsApp/SMS integration
twilio>=8.10.0,<9.0.0

# AI/ML integration
openai>=1.3.0,<2.0.0

# Configuration and utilities
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.6,<1.0.0
requests>=2.31.0,<3.0.0

# Security and rate limiting
slowapi>=0.1.9,<1.0.0 